fixed describe_user call in user property of QuickSight helper
The describe_user call in the user property splits the AWS ARN on
'/' and gets the last 2 values from the string. This might cause
issue when the ARN is an IAM User ARN (As described in #174). This
commit fixes this behaviour by removing the initial
'arn:aws:iam...' string instead.

Signed-off-by: Sampath Balivada <balivadask2000@gmail.com>